requirements.txt
pytest~=7.2.2
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
PyYAML~=6.0
requests~=2.28.2
Flask~=2.2.3
jsonpath~=0.82
rsa~=4.9
mysqlclient~=2.1.1